requirements.txt
Django>=4.2,<5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv

